Modify hub port to be configurable in tests
To prevent test collisions and make things easier for the user, this
commit introduces a GP_UPGRADE_HUB_PORT environment variable that can be
set to point the CLI to a hub on a port other than the default 7257, and
the hubs in the tests now start on a random port.

The hub started in gp_upgrade still uses the default port, as making
that configurable will require further effort.

Co-authored-by: Jeremy Alvis <jalvis@pivotal.io>
Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>