WIP: EOD commit
The bug is in not using switching to the newly minted target cluster
before loading the config and then doing a target.Commit().

Our next move is to pull the async bit into a helper that can be tested
in isolation.  We were struggling against the desire not to add
something else to the hub, but InitCluster currently is deeply
integrated to the hub, and that is causing challenges.

Co-authored-by: Jim Doty <jdoty@pivotal.io>
Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>